Refactor product detail feature: reorganize repository and entity structure, and remove unused files
