+++ lib/src/feature/productDetail/domain/entity/product_detail_entity.dart
