+++ lib/src/feature/productDetail/domain/usecase/produc_detail_usecase.dart
@@ -1,5 +1,5 @@
-import 'package:quickdrop/src/domain/entity/product_detail_entity.dart';
-import 'package:quickdrop/src/domain/repository/product_detail_repository.dart';
+import 'package:quickdrop/src/feature/productDetail/domain/entity/product_detail_entity.dart';
+import 'package:quickdrop/src/feature/productDetail/domain/repository/product_detail_repository.dart';
 
 class ProducDetailUsecase {
   ProducDetailUsecase({
