+++ lib/src/feature/productDetail/presentation/cubit/product_state.dart
