+++ lib/src/feature/productDetail/presentation/widgets/body/product_body.dart
@@ -2,11 +2,11 @@ import 'package:extensions/extensions.dart';
 import 'package:flutter/material.dart';
 import 'package:quickdrop/src/core/constants/constants.dart';
 import 'package:quickdrop/src/core/functions/price_formatter.dart';
-import 'package:quickdrop/src/domain/entity/product_detail_entity.dart';
-import 'package:quickdrop/src/presentation/product/widgets/buttons/buy_buttons.dart';
-import 'package:quickdrop/src/presentation/product/widgets/relatedProducts/related_product.dart';
-import 'package:quickdrop/src/presentation/product/widgets/tags/tags.dart';
-import 'package:quickdrop/src/presentation/product/widgets/vendor/seller.dart';
+import 'package:quickdrop/src/feature/productDetail/domain/entity/product_detail_entity.dart';
+import 'package:quickdrop/src/feature/productDetail/presentation/widgets/buttons/buy_buttons.dart';
+import 'package:quickdrop/src/feature/productDetail/presentation/widgets/relatedProducts/related_product.dart';
+import 'package:quickdrop/src/feature/productDetail/presentation/widgets/tags/tags.dart';
+import 'package:quickdrop/src/feature/productDetail/presentation/widgets/vendor/seller.dart';
 
 class ProductBody extends StatefulWidget {
   const ProductBody({
