+++ lib/src/feature/productDetail/presentation/widgets/loading/loading_product_shimmer.dart
