+++ lib/src/feature/productDetail/presentation/widgets/productHeader/product_header.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:quickdrop/src/core/constants/constants.dart';
-import 'package:quickdrop/src/presentation/product/productCubit/product_cubit.dart';
-import 'package:quickdrop/src/presentation/product/widgets/productHeader/widgets/carrousell.dart';
-import 'package:quickdrop/src/presentation/product/widgets/productHeader/widgets/title_header.dart';
+import 'package:quickdrop/src/feature/productDetail/presentation/cubit/product_cubit.dart';
+import 'package:quickdrop/src/feature/productDetail/presentation/widgets/productHeader/widgets/carrousell.dart';
+import 'package:quickdrop/src/feature/productDetail/presentation/widgets/productHeader/widgets/title_header.dart';
 import 'package:quickdrop/src/presentation/widgets/clipper_radius_images.dart';
 import 'package:quickdrop/src/presentation/widgets/header_buton.dart';
 import 'package:quickdrop/src/presentation/widgets/image_loader.dart';
