+++ lib/src/feature/productDetail/presentation/widgets/productHeader/widgets/full_screen_image.dart
