+++ lib/src/feature/productDetail/presentation/widgets/quantity/quantity.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:quickdrop/src/core/constants/constants.dart';
 import 'package:quickdrop/src/presentation/pruchase/purchaseCubit/purchase_cubit.dart';
-import 'package:quickdrop/src/presentation/product/widgets/quantity/widgets/quantity_button.dart';
-import 'package:quickdrop/src/presentation/product/widgets/quantity/widgets/quantity_text_field.dart';
+import 'package:quickdrop/src/feature/productDetail/presentation/widgets/quantity/widgets/quantity_button.dart';
+import 'package:quickdrop/src/feature/productDetail/presentation/widgets/quantity/widgets/quantity_text_field.dart';
 
 class Quantity extends StatefulWidget {
   const Quantity({super.key});
