+++ lib/src/presentation/financial_information/cubit/financial_information_cubit.dart
@@ -1,6 +1,6 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:quickdrop/src/domain/entity/product_detail_entity.dart';
+import 'package:quickdrop/src/feature/productDetail/domain/entity/product_detail_entity.dart';
 import 'package:quickdrop/src/domain/usecase/financial_information_usecase.dart';
 
 part 'financial_information_state.dart';
